Minor updates to folder structure and some changes to datawarehouse.py folder

+++ datawarehouse.sql
@@ -143,5 +143,5 @@ DROP TABLE IF EXISTS property_info;
 CREATE TABLE property_info(
     PROPTYPE text,
     NBHDNAME text,
-    ASSESSMENT integer
+    ASSESSMENT float
 );
